% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_quiz_questionviz_fct_graphs.R
\name{chart_multiplechoise_multiple}
\alias{chart_multiplechoise_multiple}
\title{Barplot for multiple choice question}
\usage{
chart_multiplechoise_multiple(quiz, question)
}
\description{
Barplot for multiple choice question
}
