% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_quiz_questionviz_fct_graphs.R
\name{chart_multiplechoise_single}
\alias{chart_multiplechoise_single}
\title{Barplot for multiple choice question - single answer}
\usage{
chart_multiplechoise_single(
  quiz,
  question,
  correct_answer = "",
  arrange_by_frequency = TRUE
)
}
\description{
Barplot for multiple choice question - single answer
}
