% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_quiz_questionviz_fct_graphs.R
\name{chart_prediction}
\alias{chart_prediction}
\title{Histogram for prediction questions}
\usage{
chart_prediction(
  quiz,
  question,
  correct_answer = NA,
  arrange_by_frequency = TRUE
)
}
\description{
Histogram for prediction questions
}
