% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_elmer.R
\name{get_ai_temperature}
\alias{get_ai_temperature}
\title{Get AI temperature configuration}
\usage{
get_ai_temperature()
}
\value{
Numeric value for temperature
}
\description{
Get the AI temperature from the golem configuration file
}
