% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_driveaccess.R
\name{get_filter_sheet}
\alias{get_filter_sheet}
\title{Get Filter Sheet Data}
\usage{
get_filter_sheet(course_directory, filter_roster, filter_sheet)
}
\arguments{
\item{course_directory}{A dribble of files of the selected course's directory}

\item{filter_roster}{A string of the Roster's filename (same file as filter sheet)}

\item{filter_sheet}{A string of the sheetname of the Filter spreadsheet}
}
\value{
A tibble of the Filter sheet data or `NULL` if fail.
}
\description{
Load filter sheet data from the same Google Sheet file as the roster.
Columns A, B, C are name_canvas, sis_canvas, email_canvas. Columns D onwards are filter variables.
}
