% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_driveaccess.R
\name{get_filter_variables}
\alias{get_filter_variables}
\title{Get Filter Variables from Filter Sheet}
\usage{
get_filter_variables(
  course_directory,
  filter_roster,
  filter_sheet,
  return_named_list = FALSE
)
}
\arguments{
\item{course_directory}{A dribble of files of the selected course's directory}

\item{filter_roster}{A string of the Roster's filename (same file as filter sheet)}

\item{filter_sheet}{A string of the sheetname of the Filter spreadsheet}

\item{return_named_list}{A logical indicating whether to return a named list with display names}
}
\value{
A character vector of filter variable names or named list with display names
}
\description{
Get list of filter variables (columns D onwards) from the filter sheet
}
