% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_driveaccess.R
\name{get_roster}
\alias{get_roster}
\title{Get Roster Sheet Processed}
\usage{
get_roster(course_directory, filter_roster, filter_roster_sheet)
}
\arguments{
\item{course_directory}{A dribble of files of the selected course's directory}

\item{filter_roster}{A string of the Roster's filename}

\item{filter_roster_sheet}{A string of the sheetname of the Roster spreadsheet}
}
\value{
A tibble of the Roster processed or `NULL` if fail.
}
\description{
Load Roster Spreadsheet and cleans it.
}
